% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.broadbandnow.broadband.prices-package.R
\docType{package}
\name{dc.broadbandnow.broadband.prices-package}
\alias{dc.broadbandnow.broadband.prices}
\alias{dc.broadbandnow.broadband.prices-package}
\title{dc.broadbandnow.broadband.prices: National Capital Region Internet Prices}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
