% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.broadbandnow.broadband.prices-package.R
\docType{package}
\name{dc.broadbandnow.broadband.prices-package}
\alias{dc.broadbandnow.broadband.prices}
\alias{dc.broadbandnow.broadband.prices-package}
\title{dc.broadbandnow.broadband.prices: National Capital Region Internet Prices}
\description{
Allows user to easily get broadband prices data from the SDAD Data Commons. Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Zachary Alerte \email{rsu8zv@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
